+++ datatest/src/commonMain/kotlin/at/asitplus/test/Intermediate.map.kt
@@ -12,7 +12,6 @@ import de.infix.testBalloon.framework.core.TestSuiteScope
  * @param map Map of suite names to test data
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
  * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  */
@@ -20,13 +19,11 @@ fun <Data> TestSuiteScope.withData(
     map: Map<String, Data>,
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength!!,
-    displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength!!,
     prefix: String = "",
     testConfig: TestConfig = TestConfig,
 ) = ConfiguredDataTestScope<Data>(
     compact,
     maxLength,
-    displayNameMaxLength = displayNameMaxLength,
     prefix = prefix,
     this,
     map.asSequence().map { (k, v) -> k to v },
@@ -41,15 +38,14 @@ fun <Data> TestSuiteScope.withData(
  * @param testConfig Optional test configuration
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
  * @param prefix an optional prefix to add to the test name
  * @param action Test suite configuration action for each map value
  */
+@Deprecated("will be removed in 0.9.0", ReplaceWith("withData(map, compact, maxLength, prefix, testConfig) - {}"))
 fun <Data> TestSuiteScope.withDataSuites(
     map: Map<String, Data>,
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength!!,
-    displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength!!,
     prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: TestSuiteScope.(Data) -> Unit
@@ -57,7 +53,6 @@ fun <Data> TestSuiteScope.withDataSuites(
     map.map { (k, v) -> k to v }.asSequence(),
     compact,
     maxLength,
-    displayNameMaxLength = displayNameMaxLength,
     prefix = prefix,
     testConfig,
     action