+++ datatest/src/commonMain/kotlin/at/asitplus/test/Terminal.map.kt
@@ -13,7 +13,6 @@ import de.infix.testBalloon.framework.core.TestSuiteScope
  * @param map Map of test names to test data
  * @param compact If true, only a single test element is created and the class name of the data parameter is used as test name
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
  * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  * @param action Test action to execute for each map value
@@ -22,7 +21,6 @@ fun <Data> TestSuiteScope.withData(
     map: Map<String, Data>,
     compact: Boolean = DataTest.compactByDefault,
     maxLength: Int = DataTest.defaultTestNameMaxLength!!,
-    displayNameMaxLength: Int = DataTest.defaultDisplayNameMaxLength!!,
     prefix: String = "",
     testConfig: TestConfig = TestConfig,
     action: suspend Test.ExecutionScope.(Data) -> Unit
@@ -31,7 +29,6 @@ fun <Data> TestSuiteScope.withData(
     testConfig,
     compact,
     maxLength,
-    displayNameMaxLength,
     prefix,
     action
 )