+++ datatest/src/commonTest/kotlin/DataTest.kt
@@ -1,10 +1,9 @@
 import at.asitplus.testballoon.withData
-import at.asitplus.testballoon.withDataSuites
 import de.infix.testBalloon.framework.core.testSuite
 import io.kotest.matchers.shouldBe
 
 val aDataDrivenSuite by testSuite {
-    withDataSuites("foo" to 13) {
+    withData("foo" to 13) - {
         test("bar") {
             it shouldBe 13
         }
@@ -15,7 +14,7 @@ val aDataDrivenSuite by testSuite {
 
 
 
-    withDataSuites(null, null, 1, 2, 3, 4, compact = true) { number ->
+    withData(null, null, 1, 2, 3, 4, compact = true) - { number ->
         withData(number.toString(), "one", null, null, null, "two", "three", "four", compact = true) { word ->
             number shouldBe number
             word shouldBe "three"