+++ freespec/src/commonMain/kotlin/at/asitplus/test/FreeSpec.kt
@@ -18,18 +18,7 @@ object FreeSpec {
      * This property's getter will never return null, but fall back to [TestBalloonAddons.defaultTestNameMaxLength].
      */
     var defaultTestNameMaxLength: Int? = null
-        get() = field?:TestBalloonAddons.defaultTestNameMaxLength
-
-    /**
-     * The default maximum length of test element display names (not test name).
-     * Defaults to [TestBalloonAddons.defaultDisplayNameMaxLength], but setting it here will take precedence.
-     * * `-1` means no truncation.
-     * * `null` means it will again fall back to [TestBalloonAddons.defaultDisplayNameMaxLength]
-     *
-     * This property's getter will never return null, but fall back to [TestBalloonAddons.defaultDisplayNameMaxLength].
-     */
-    var defaultDisplayNameMaxLength: Int? = null
-        get() = field?:TestBalloonAddons.defaultDisplayNameMaxLength
+        get() = field ?: TestBalloonAddons.defaultTestNameMaxLength
 
 }
 
@@ -39,16 +28,12 @@ context(suite: TestSuiteScope)
  *
  * @param testConfig Optional test configuration
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
- * @property displayName optional display name override
  * @param nested The test body to execute.
  */
 @Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE")
 @kotlin.internal.LowPriorityInOverloadResolution
 operator fun String.invoke(
-    displayName: String = this,
     maxLength: Int = FreeSpec.defaultTestNameMaxLength!!,
-    displayNameMaxLength: Int = FreeSpec.defaultDisplayNameMaxLength!!,
     testConfig: TestConfig = TestConfig,
     nested: suspend Test.ExecutionScope.() -> Unit
 ) {
@@ -56,10 +41,9 @@ operator fun String.invoke(
         val truncatedName = freeSpecName(this@invoke).truncated(maxLength)
         testSuiteInScope.checkPathLenIncluding(truncatedName)
         test(
-            truncatedName,
-            displayName = (displayName.truncated(displayNameMaxLength)),
+            name = truncatedName,
             testConfig = testConfig.disableByName(this@invoke),
-            nested
+            action = nested
         )
     }
 }
@@ -70,16 +54,12 @@ operator fun String.invoke(
  *
  * @property parent The parent test suite
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
  * @property testName The name of the suite
- * @property displayName optional display name override
  * @property config The configuration for the suite
  */
 data class ConfiguredSuite(
     val parent: TestSuiteScope,
     val maxLength: Int = FreeSpec.defaultTestNameMaxLength!!,
-    val displayNameMaxLength: Int = FreeSpec.defaultDisplayNameMaxLength!!,
-    val displayName: String,
     val testName: String,
     val config: TestConfig
 ) {
@@ -94,8 +74,7 @@ data class ConfiguredSuite(
             testSuiteInScope.checkPathLenIncluding(truncatedName)
             testSuite(
                 truncatedName,
-                displayName = (displayName.truncated(displayNameMaxLength)),
-                testConfig = config.disableByName(displayName),
+                testConfig = config.disableByName(testName),
                 content = suiteBody
             )
         }
@@ -110,17 +89,13 @@ context(suite: TestSuiteScope)
  *
  * @param testConfig Optional test configuration
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
- * @param displayName Optional display name override
  * @return A new [ConfiguredSuite] instance.
  */
 operator fun String.invoke(
-    displayName: String = this,
     maxLength: Int = FreeSpec.defaultTestNameMaxLength!!,
-    displayNameMaxLength: Int = FreeSpec.defaultDisplayNameMaxLength!!,
     testConfig: TestConfig = TestConfig
 ) =
-    ConfiguredSuite(suite, maxLength, displayNameMaxLength, displayName, this, testConfig)
+    ConfiguredSuite(suite, maxLength, this, testConfig)
 
 context(suite: TestSuiteScope)
 /**
@@ -134,7 +109,6 @@ infix operator fun String.minus(suiteBody: TestSuiteScope.() -> Unit) =
         testSuiteInScope.checkPathLenIncluding(truncatedName)
         testSuite(
             name = truncatedName,
-            displayName = (freeSpecName(this@minus).truncated(FreeSpec.defaultDisplayNameMaxLength!!)),
             testConfig = TestConfig.disableByName(this@minus),
             content = suiteBody
         )