+++ property/src/commonMain/kotlin/at/asitplus/test/Property.terminal.kt
@@ -15,7 +15,6 @@ import io.kotest.property.PropertyTesting
  * @param genA Generator for test values
  * @param compact whether to compact all generated child test elements into one
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
  * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  * @param content Test execution block receiving generated values
@@ -24,16 +23,14 @@ fun <Value> TestSuiteScope.checkAll(
     genA: Gen<Value>,
     compact: Boolean = PropertyTest.compactByDefault,
     maxLength: Int = PropertyTest.defaultTestNameMaxLength!!,
-    displayNameMaxLength: Int = PropertyTest.defaultDisplayNameMaxLength!!,
     prefix: String = "",
     testConfig: TestConfig = TestConfig,
     content: suspend context(PropertyContext) Test.ExecutionScope.(Value) -> Unit
 ) = checkAll(
-    PropertyTesting.defaultIterationCount,
+    iterations = PropertyTest.defaultIterationCount,
     genA,
     compact,
     maxLength,
-    displayNameMaxLength,
     prefix,
     testConfig,
     content
@@ -46,7 +43,6 @@ fun <Value> TestSuiteScope.checkAll(
  * @param genA Generator for test values
  * @param compact whether to compact all generated child test elements into one
  * @param maxLength maximum length of test element name (not display name)
- * @param displayNameMaxLength maximum length of test element **display name**
  * @param prefix an optional prefix to add to the test name
  * @param testConfig Optional test configuration
  * @param content Test execution block receiving generated values
@@ -56,7 +52,6 @@ fun <Value> TestSuiteScope.checkAll(
     genA: Gen<Value>,
     compact: Boolean = PropertyTest.compactByDefault,
     maxLength: Int = PropertyTest.defaultTestNameMaxLength!!,
-    displayNameMaxLength: Int = PropertyTest.defaultDisplayNameMaxLength!!,
     prefix: String = "",
     testConfig: TestConfig = TestConfig,
     content: suspend context(PropertyContext) Test.ExecutionScope.(Value) -> Unit
@@ -65,7 +60,6 @@ fun <Value> TestSuiteScope.checkAll(
     genA,
     compact,
     maxLength,
-    displayNameMaxLength,
     prefix,
     testConfig,
 ) { content(it) }